Phew, just needed to freeze the correct requirements. This could have gone a lot worse!
For instance, if the Django version were any older it wouldn't have worked
with Python 3.9, same for compiling Pillow. But I couldn't upgrade it because
this projects is stuck on an old version of SimpleCMS, which I also couldn't
have fixed in case of a Python 3.9 incompatibility.

In the end it all worked out and this configuration should last until the
next Debian release. Then, you're fucked.